Refactor: Default to English names if system language is not Spanish
This commit modifies the behavior of how country names are displayed in several parts of the application.

Previously, country names were displayed in English if the device's locale was set to English ("en"), and in Spanish otherwise.

Now, country names will be displayed in Spanish if the device's locale is Spanish ("es"). For all other locales (including English), the English names will be used.

This change affects:
- `GuessCapitalScreen.kt`: Option buttons now show Spanish names if locale is "es", otherwise English.
- `GuessFlagScreen.kt`: Option buttons now show Spanish names if locale is "es", otherwise English.
- `CountriesInfoScreen.kt`:
    - The country list is sorted by Spanish names if locale is "es", otherwise by English names.
    - Displayed country names (in search suggestions and details) use Spanish if locale is "es", otherwise English.